+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 
 - Fixed `impl From<RiferimentoNISECI> for Vec<SpecieNISECI>` to use `map_ids` field
+- Bump `chrono` to `0.4.45`
 
 ## [0.1.6] - 2026-05-15