+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed `impl From<RiferimentoNISECI> for Vec<SpecieNISECI>` to use `map_ids` field
 - Bump `chrono` to `0.4.45`
+- Bump `serde_json` to `1.0.150`
 
 ## [0.1.6] - 2026-05-15